ner_app/requirements.txt
spacy>=3.7.0
pyyaml>=6.0
